requirements.txt
flask==1.1.1
pillow==6.2.0
jwt==1.0.0
requests==2.22.0
markdown==3.3